frameworks/base:(Hack)Print the content displayed with WebView
This hack allows us to print content that displayed with
the loadDataWithBaseURL method of WebViewClassic.java
when ro.debug.loadDataWithBaseURL is set to true.
This allows us to capture the contents of WebView which will help
us capture benchmark results for thirdparty apps
which use loadDataWithBaseURL of WebViewClassic to display results.

Change-Id: Id8e61e072b867501ef2e0e8eb2f6fb30b13206b4
Signed-off-by: Yongqin Liu <yongqin.liu@linaro.org>